Makefile: Build for stdmem.o. Build and run pascal array tests.
stdmem.c:
We need stdmonads.h, include it.
libandria4_mallocwrapper_FAILURE() needs to take the .val member of it's arg.
Rename the argument to libandria4_mallocwrapper_FAILURE() to fix double-usage bug.
Do the same two things for the *realloc*() version of the macro.
The call to libandria4_errno_2result() in *_deallocwrapper() needs a semicolon to end the statement.

pascalarraytest.c:
visit_checker() doesn't have a "parr" variable, replace with ref.
Add a comma after the "   False failure"... string in BUILDFAIL() in main().
Fix the escape sequences in the "   Content failure"... string in TEST_CONTENT() in main().
Fix the missing ( (char[]){ GROWN_TEXT } ) at line 290.
Fix the "parr instead of parr1" typo at line 326 in main().

Signed-off-by: Jam555 <3349478+jam555@users.noreply.github.com>